# Optimize reflection of F# types