While compiling expression trees to `Func`s for faster execution is probably an overkill for one-off reflection functions, I reckon this extra step is worth it for their precomputed counterparts, which tend to be used in performance-critical scenarios such as serialization.

If there's interest, I can similarly improve other `PreCompute` functions.