@abelbraaksma 2020-07-18
Are you sure `Compile` method is available in .NET Core? I don't see it mentioned here: https://docs.microsoft.com/en-us/dotnet/api/system.data.objects.compiledquery.compile?view=netframework-4.8#System_Data_Objects_CompiledQuery_Compile__4_System_Linq_Expressions_Expression_System_Func___0___1___2___3___, at the bottom it only lists .NET Framework.