@abelbraaksma 2020-07-18
This will greatly improve repeated calls, but I wonder where the threshold is, because compiling is quite expensive. I mean, is it beneficial from 5 calls up, or 500 calls?

I think this is a great improvement, but we should probably know how big the performance improvement is, and where it starts to improve. Did you test with BDN? 

Would it be possible to have the best of both, for instance by calling it the old way, and lazily compile in a different thread (no idea this is even feasible). Once compilation is done, subsequent calls come from the compiled one.