@kerams 2020-07-18
```fsharp
type Record = {
    A: int
    B: int
    C: string
    D: string
    E: unit }

let compilePropGetterFunc (prop: PropertyInfo) =
    let param = Expression.Parameter (typeof<obj>, "param")

    let expr =
        Expression.Lambda<Func<obj, obj>> (
            Expression.Convert (
                Expression.Property (
                    Expression.Convert (param, prop.DeclaringType),
                    prop),
                typeof<obj>),
            param)
    expr.Compile ()

[<MemoryDiagnoser>]
type Test () =
    let before = FSharpValue.PreComputeRecordReader typeof<Record>
    let after  =
        let fs =
            typeof<Record>.GetProperties (BindingFlags.Instance ||| BindingFlags.Public)
            |> Array.map compilePropGetterFunc
    
        fun (o: obj) -> fs |> Array.map (fun f -> f.Invoke o)
    
    let record = { A = 1; B = 2; C = "3"; D = "4"; E = () }

    [<Benchmark>]
    member _.Before () =
        for i in 1 .. 1000 do
            before record |> ignore

    [<Benchmark>]
    member _.After () =
        for i in 1 .. 1000 do
            after record |> ignore

BenchmarkRunner.Run<Test> () |> ignore
```

| Method |      Mean |    Error |   StdDev | Ratio |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |----------:|---------:|---------:|------:|--------:|------:|------:|----------:|
| Before | 537.71 us | 5.888 us | 5.219 us |  1.00 | 12.6953 |     - |     - | 109.38 KB |
|  After |  33.57 us | 0.511 us | 0.427 us |  0.06 | 13.3667 |     - |     - | 109.38 KB |