@abelbraaksma 2020-07-18
If I understand the PR correctly, this pre-compiles, then caches access to members of records when `PreComputeRecordReader` is used, right? And you deliberately didn't do it for `GetRecordFields`. Since that method already has the word `compute` in it, it kinda makes sense. But I agree that it would be even better to be have this for more functions in `reflect.fs`. 

I am, however, a little worried about the overhead. I don't know in what contexts this code is usually used (well, in reflection), and if we can ascertain somehow that the threshold of 350+ calls is reached. 

An alternative would be do add functions that have `Compile` in the name, so that users can choose. Something like `PreCompileRecordReader`, `GetCompiledRecordFields`. 

Yet another way is perhaps how _dynamics_ works in C#, which caches the invoked member for future access, though I'm not sure if it compiles. I mean, I think in C# you get the `MethodInfo`, and a delegate is created, which is much cheaper than compiling a LINQ tree.