@kerams 2020-07-18
>If I understand the PR correctly, this pre-compiles, then caches access to members of records when PreComputeRecordReader is used, right?

The property accessors are embedded in the returned function closure. You can refer to it as caching, but if you call the precompute method again with the same type, everything is compiled anew and you get a different closure back.

>you deliberately didn't do it for GetRecordFields

Yes, that's the "one-off" variant I talked about in the OP. There isn't any sort of caching involved and each call looks up `PropertyInfo` of each field of the record type. See `PropertyLookupEveryTime` method in the benchmark.

>and if we can ascertain somehow that the threshold of 350+ calls is reached

Unless I misunderstood something, the caller is the one that needs to know how often they're going to need to read record fields and it is their responsibility to choose the appropriate API/approach.

>An alternative would be do add functions that have Compile in the name, so that users can choose. Something like PreCompileRecordReader, GetCompiledRecordFields.

Sure, but it would also be fantastic if users could automatically reap the benefits of this change by simply updating to a new version of .NET/FSharp.Core. As we have established though, this does introduce additional overhead, so I'll let the powers that be decide whether or not a new set of methods is required.

>and a delegate is created, which is much cheaper than compiling a LINQ tree

I think `Delegate.CreateDelegate` returns a delegate for a specific instance if used with instance methods. It's also quite a bit faster than plain `Invoke`, but nowhere near as fast the compiled `Func`, which technically isn't even reflection anymore. Compare `Direct` and `After2` benchmarks. The only overhead stems from the need to allocate an array to return the results and boxing of value types.